fix: document not defined in onboarding
